[CCUBE-303][RL] Rename components and barrel export files
