+++ src/shared/internal-calendar/fixed-range/fixed-range-calendar-day-view.tsx
@@ -2,11 +2,7 @@ import dayjs, { Dayjs } from "dayjs";
 import { useMemo, useState } from "react";
 import { Text } from "../../../text/text";
 import { CalendarHelper } from "../../../util/calendar-helper";
-import {
-    HeaderCell,
-    RowDayCell,
-    Wrapper,
-} from "../day-of-month/calendar-day-view.style";
+import { HeaderCell, RowDayCell, Wrapper } from "../standard";
 import { CommonCalendarProps } from "../types";
 import { FixedRangeDayCell } from "./fixed-range-cell";
