+++ src/shared/internal-calendar/fixed-range/index.ts
@@ -1,0 +1,1 @@
+export * from "./fixed-range-calendar-day-view";
