+++ src/shared/internal-calendar/week/index.ts
@@ -1,0 +1,1 @@
+export * from "./week-calendar-day-view";
